review updates and cleanup
