+++ src/module.cpp
@@ -1468,7 +1468,7 @@ std::vector<module_ref> module::get_sub_modules(bool shallow) const
     return vec_modules;
 }
 
-module module::with_static_shapes(const std::vector<shape>& input_shapes)
+module module::with_static_shapes(const std::unordered_map<std::string, shape>& input_shapes)
 {
     // This routine creates a new module with the same instructions but with different input shapes.
     // The sequence of instructions (operators and interconnectivity) is copied, but all input
@@ -1478,64 +1478,23 @@ module module::with_static_shapes(const std::vector<shape>& input_shapes)
     auto param_names = this->get_parameter_names();
     assert(param_names.size() == input_shapes.size());
 
-    // Make a mapping from the parameter names to the new shapes.
-    std::unordered_map<std::string, shape> shape_map;
-    for(std::size_t i = 0; i < param_names.size(); ++i)
-        shape_map[param_names[i]] = input_shapes[i];
-
     module new_mod;
-
     std::unordered_map<instruction_ref, instruction_ref> ins_map;
 
-    // First, create parameters with new shapes in new_mod and fill ins_map for params
+    // create parameters with new shapes in new_mod and fill ins_map for params
     for(auto ins : iterator_for(*this))
     {
         if(ins->name() == "@param")
         {
             auto pname = any_cast<builtin::param>(ins->get_operator()).parameter;
-            assert(shape_map.count(pname) > 0);
-            ins_map[ins] = new_mod.add_parameter(pname, shape_map.at(pname));
+            assert(input_shapes.count(pname) > 0);
+            ins_map[ins] = new_mod.add_parameter(pname, input_shapes.at(pname));
         }
     }
 
-    // Copy remaining instructions (except parameters) in order
-    for(auto ins : iterator_for(*this))
-    {
-        if(ins->name() == "@param")
-            continue;
-
-        // Gather new input refs for this instruction
-        std::vector<instruction_ref> new_args;
-        std::transform(ins->inputs().begin(),
-                       ins->inputs().end(),
-                       std::back_inserter(new_args),
-                       [&](auto arg) { return ins_map.at(arg); });
-
-        // Gather new module argument refs if present
-        std::vector<module_ref> new_mod_args;
-        std::transform(ins->module_inputs().begin(),
-                       ins->module_inputs().end(),
-                       std::back_inserter(new_mod_args),
-                       [&](auto modarg) { return modarg; });
-
-        instruction_ref new_ins;
-        if(ins->name() == "@literal")
-        {
-            new_ins = new_mod.add_literal(ins->get_literal());
-        }
-        else if(ins->name() == "@return")
-        {
-            new_ins = new_mod.add_return(new_args);
-        }
-        else
-        {
-            if(new_mod_args.empty())
-                new_ins = new_mod.add_instruction(ins->get_operator(), new_args);
-            else
-                new_ins = new_mod.add_instruction(ins->get_operator(), new_args, new_mod_args);
-        }
-        ins_map[ins] = new_ins;
-    }
+    // Copy remaining instructions in order
+    auto ret = new_mod.insert_instructions(new_mod.end(), this, &ins_map);
+    new_mod.add_return(ret);
 
     return new_mod;
 }
